app/core/config.py
from typing import List

from dotenv import load_dotenv
from pydantic_settings import BaseSettings

load_dotenv()


class Settings(BaseSettings):
    PROJECT_NAME: str = "Todo List API"
    API_V1_STR: str = "/api/v1"

    DATABASE_URL: str

    SECRET_KEY: str
    ACCESS_TOKEN_EXPIRE_MINUTES: int = 30
    BACKEND_CORS_ORIGINS: List[str] = ["*"]

    OTEL_EXPORTER_OTLP_ENDPOINT: str

    class Config:
        case_sensitive = True


settings = Settings()  # type: ignore
