requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
alembic
psycopg2-binary

passlib[bcrypt]
python-jose[cryptography]

pydantic
pydantic-settings

prometheus-fastapi-instrumentator

opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-logging
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-psycopg2
opentelemetry-exporter-otlp-proto-http

python-dotenv
python-multipart
